moved validatePlan()
from MoveGroupInterface (client side)
to MoveGroup's ExecutionServiceCapability (server side)

This checks validity of plan no matter which client called.
Placing it in MoveGroupContext will allow re-use by other capabilities.
