+++ visualization/motion_planning_rviz_plugin/src/motion_planning_frame_planning.cpp
@@ -139,9 +139,7 @@ void MotionPlanningFrame::computeExecuteButtonClicked()
   if (move_group_ && current_plan_)
   {
     ui_->stop_button->setEnabled(true); // enable stopping
-    bool success =
-      move_group_->validatePlan(*current_plan_) &&
-      move_group_->execute(*current_plan_);
+    bool success = move_group_->execute(*current_plan_);
     onFinishedExecution(success);
   }
 }
